requirements.txt
PyJWT
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0